+++ app/models/article.py
@@ -1,0 +1,15 @@
+from tortoise import fields
+from tortoise.models import Model
+
+class Article(Model):
+
+    id = fields.IntField(pk=True)
+
+    title = fields.TextField()
+    content = fields.TextField()
+
+    created_at = fields.DatetimeField(auto_now_add=True)
+    updated_at = fields.DatetimeField(auto_now=True)
+
+    def __str__(self):
+        return self.title
